refactor: is owner of post middleware

+++ routes/web.php
@@ -3,6 +3,7 @@
 use App\Http\Controllers\HomeController;
 use App\Http\Controllers\PostController;
 use App\Http\Controllers\UserController;
+use App\Http\Middleware\IsPostOwner;
 use Illuminate\Support\Facades\Route;
 
 // Home routes
@@ -17,7 +18,7 @@ Route::post('/login', [UserController::class, 'login']);
 
 // Post routes
 Route::get('/post', [PostController::class,'createPostView']);
+Route::get('/post/{post}/edit', [PostController::class, 'editPostView'])->middleware(IsPostOwner::class);
 Route::post('/post', [PostController::class, 'createPost']);
-Route::get('/post/{post}/edit', [PostController::class, 'editPostView']);
-Route::put('/post/{post}', [PostController::class, 'editPost']);
-Route::delete('/post/{post}', [PostController::class, 'deletePost']);
+Route::put('/post/{post}', [PostController::class, 'editPost'])->middleware(IsPostOwner::class);
+Route::delete('/post/{post}', [PostController::class, 'deletePost'])->middleware(IsPostOwner::class);
